fix compilation error

+++ stringify/float.go
@@ -7,5 +7,5 @@ func Float64(f float64) string {
 }
 
 func Float32(f float32) string {
-	return strconv.FormatFloat(f, 'E', -1, 32)
+	return strconv.FormatFloat(float64(f), 'E', -1, 32)
 }
